11: Finished generic datatable
