--- src/lib/components/custom/dataTable/(components)/data-table.svelte
@@ -1,211 +1,0 @@
-<script lang="ts">
-	import { get, readable } from 'svelte/store';
-	import { Render, Subscribe, createRender, createTable } from 'svelte-headless-table';
-	import * as Table from '$lib/components/ui/table/';
-	import {
-		addColumnFilters,
-		addHiddenColumns,
-		addPagination,
-		addSelectedRows,
-		addSortBy,
-		addTableFilter
-	} from 'svelte-headless-table/plugins';
-	import {
-		DataTableCheckbox,
-		DataTableDescriptionCell,
-		DataTableExpLevelCell,
-		DataTableRowActions,
-		DataTableModalityCell,
-		DataTableColumnHeader,
-		DataTableToolbar,
-		DataTablePagination
-	} from '.';
-	import type { Program } from '../(data)/schemas';
-
-	export let data: Program[];
-
-	const table = createTable(readable(data), {
-		select: addSelectedRows(),
-		sort: addSortBy({
-			toggleOrder: ['asc', 'desc']
-		}),
-		page: addPagination(),
-		filter: addTableFilter({
-			fn: ({ filterValue, value }) => {
-				return value.toLowerCase().includes(filterValue.toLowerCase());
-			}
-		}),
-		colFilter: addColumnFilters(),
-		hide: addHiddenColumns()
-	});
-
-	const columns = table.createColumns([
-		table.display({
-			id: 'select',
-			header: (_, { pluginStates }) => {
-				const { allPageRowsSelected } = pluginStates.select;
-				return createRender(DataTableCheckbox, {
-					checked: allPageRowsSelected,
-					'aria-label': 'Select all'
-				});
-			},
-			cell: ({ row }, { pluginStates }) => {
-				const { getRowState } = pluginStates.select;
-				const { isSelected } = getRowState(row);
-				return createRender(DataTableCheckbox, {
-					checked: isSelected,
-					'aria-label': 'Select row',
-					class: 'translate-y-[2px]'
-				});
-			},
-			plugins: {
-				sort: {
-					disable: true
-				}
-			}
-		}),
-		table.column({
-			accessor: 'name',
-			header: 'Name',
-			id: 'name',
-			plugins: {
-				sort: {
-					disable: true
-				}
-			}
-		}),
-		table.column({
-			accessor: 'description',
-			header: 'Description',
-			id: 'description',
-			cell: ({ value, row }) => {
-				if (row.isData()) {
-					return createRender(DataTableDescriptionCell, {
-						value
-					});
-				}
-				return value;
-			}
-		}),
-		table.column({
-			accessor: 'modality',
-			header: 'Modality',
-			id: 'modality',
-			cell: ({ value }) => {
-				return createRender(DataTableModalityCell, {
-					value
-				});
-			},
-			plugins: {
-				colFilter: {
-					fn: ({ filterValue, value }) => {
-						if (filterValue.length === 0) return true;
-						if (!Array.isArray(filterValue) || typeof value !== 'string') return true;
-						return filterValue.some((filter) => {
-							return value.includes(filter);
-						});
-					},
-					initialFilterValue: [],
-					render: ({ filterValue }) => {
-						return get(filterValue);
-					}
-				}
-			}
-		}),
-		table.column({
-			accessor: 'experienceLevel',
-			id: 'experiencelevel',
-			header: 'Experience Level',
-			cell: ({ value }) => {
-				return createRender(DataTableExpLevelCell, {
-					value
-				});
-			},
-			plugins: {
-				colFilter: {
-					fn: ({ filterValue, value }) => {
-						if (filterValue.length === 0) return true;
-						if (!Array.isArray(filterValue) || typeof value !== 'string') return true;
-
-						return filterValue.some((filter) => {
-							return value.includes(filter);
-						});
-					},
-					initialFilterValue: [],
-					render: ({ filterValue }) => {
-						return get(filterValue);
-					}
-				}
-			}
-		}),
-		table.display({
-			id: 'actions',
-			header: () => {
-				return '';
-			},
-			cell: ({ row }) => {
-				if (row.isData() && row.original) {
-					return createRender(DataTableRowActions, {
-						row: row.original
-					});
-				}
-				return '';
-			}
-		})
-	]);
-
-	const tableModel = table.createViewModel(columns);
-
-	const { headerRows, pageRows, tableAttrs, tableBodyAttrs } = tableModel;
-</script>
-
-<div class="space-y-4">
-	<DataTableToolbar {tableModel} />
-	<div class="rounded-md border">
-		<Table.Root {...$tableAttrs}>
-			<Table.Header>
-				{#each $headerRows as headerRow}
-					<Subscribe rowAttrs={headerRow.attrs()}>
-						<Table.Row>
-							{#each headerRow.cells as cell (cell.id)}
-								<Subscribe attrs={cell.attrs()} let:attrs props={cell.props()} let:props>
-									<Table.Head {...attrs}>
-										{#if cell.id !== 'select' && cell.id !== 'actions'}
-											<DataTableColumnHeader {props}
-												><Render of={cell.render()} /></DataTableColumnHeader
-											>
-										{:else}
-											<Render of={cell.render()} />
-										{/if}
-									</Table.Head>
-								</Subscribe>
-							{/each}
-						</Table.Row>
-					</Subscribe>
-				{/each}
-			</Table.Header>
-			<Table.Body {...$tableBodyAttrs}>
-				{#each $pageRows as row (row.id)}
-					<Subscribe rowAttrs={row.attrs()} let:rowAttrs>
-						<Table.Row {...rowAttrs}>
-							{#each row.cells as cell (cell.id)}
-								<Subscribe attrs={cell.attrs()} let:attrs>
-									<Table.Cell {...attrs}>
-										{#if cell.id === 'task'}
-											<div class="w-[80px]">
-												<Render of={cell.render()} />
-											</div>
-										{:else}
-											<Render of={cell.render()} />
-										{/if}
-									</Table.Cell>
-								</Subscribe>
-							{/each}
-						</Table.Row>
-					</Subscribe>
-				{/each}
-			</Table.Body>
-		</Table.Root>
-	</div>
-	<DataTablePagination {tableModel} />
-</div>
