--- src/lib/components/custom/dataTable/(components)/index.ts
@@ -1,10 +1,0 @@
-export { default as DataTableCheckbox } from './datacheckbox.svelte';
-export { default as DataTableExpLevelCell } from './cells/experienceLevel-cell.svelte';
-export { default as DataTableRowActions } from './row-actions.svelte';
-export { default as DataTableDescriptionCell } from './cells/description-cell.svelte';
-export { default as DataTableModalityCell } from './cells/modality-cell.svelte';
-export { default as DataTableColumnHeader } from './column-header.svelte';
-export { default as DataTableToolbar } from './toolbar.svelte';
-export { default as DataTablePagination } from './pagination.svelte';
-export { default as DataTableViewOptions } from './view-options.svelte';
-export { default as DataTableFacetedFilter } from './faceted-filter.svelte';
