--- src/lib/components/custom/dataTable/(data)/schemas.ts
@@ -1,10 +1,0 @@
-import { z } from 'zod';
-
-export const programSchema = z.object({
-	name: z.string(),
-	description: z.string(),
-	modality: z.string(),
-	experienceLevel: z.string()
-});
-
-export type Program = z.infer<typeof programSchema>;
