--- src/lib/components/custom/dataTable/index.ts
@@ -1,1 +1,0 @@
-export { default as DataTable } from './+page.svelte';
