+++ src/lib/components/custom/generic-datatable/index.ts
@@ -4,3 +4,5 @@ export { default as ColumnHeader } from './column-header.svelte';
 export { default as GenericDataTable } from './table.svelte';
 export { default as GenericCell } from './generic-cell.svelte';
 export { default as Pagination } from './pagination.svelte';
+export { default as ViewOptions } from './view-options.svelte';
+export { default as Toolbar } from './toolbar.svelte';
