+++ src/routes/(app)/workouts/+page.svelte
@@ -24,13 +24,12 @@
 			id: 'experienceLevel',
 			header: 'Experience Level'
 		},
-        {
+		{
 			accessor: 'muscleGroups',
 			id: 'muscleGroups',
 			header: 'Muscle Groups'
 		}
 	];
-
 </script>
 
 <div class="hidden h-full flex-1 flex-col space-y-8 p-8 md:flex">
